Move FormulaDialog
